pyproject.toml
[tool.poetry]
name = "mnlm"
version = "0.1.0"
description = ""
authors = ["Yx Jiang <2237303+yxjiang@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
openai = "^1.6.1"
pygame = "^2.5.2"
pyaudio = "^0.2.14"
tavily-python = "^0.3.0"
colorama = "^0.4.6"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
